feat: Add PDFViewerScreen component for displaying PDF files
