+++ KrishnaAcademyAPP/app/(routes)/pdfviewer.tsx
@@ -1,0 +1,51 @@
+import { useNavigation } from 'expo-router';
+import React from 'react';
+import { View, Text, StyleSheet, TouchableOpacity } from 'react-native';
+import Pdf from 'react-native-pdf';
+
+const PDFViewerScreen = () => {
+    const route = useNavigation();
+
+  const { pdfUri } = route.params;
+
+  return (
+    <View style={styles.container}>
+      <TouchableOpacity style={styles.closeButton} onPress={() => navigation.goBack()}>
+        <Text style={styles.closeButtonText}>Close</Text>
+      </TouchableOpacity>
+      {/* <Pdf
+        source={{ uri: pdfUri }}
+        onLoadComplete={(numberOfPages, filePath) => {
+          console.log(`Number of pages: ${numberOfPages}`);
+        }}
+        onPageChanged={(page, numberOfPages) => {
+          console.log(`Current page: ${page}`);
+        }}
+        onError={(error) => {
+          console.log(error);
+        }}
+        style={styles.pdf}
+      /> */}
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: 'white',
+    borderRadius: 10,
+    overflow: 'hidden',
+  },
+  closeButton: {
+    padding: 10,
+  },
+  closeButtonText: {
+    color: 'blue',
+  },
+  pdf: {
+    flex: 1,
+  },
+});
+
+export default PDFViewerScreen;
